added Services interfaces
