+++ wallet-web/src/main/java/com/isa/jjdzr/walletweb/config/WalletWebConfiguration.java
@@ -2,7 +2,9 @@ package com.isa.jjdzr.walletweb.config;
 
 import com.isa.jjdzr.walletcore.market.Market;
 import com.isa.jjdzr.walletcore.service.WalletAssetService;
+import com.isa.jjdzr.walletcore.service.WalletAssetServiceImpl;
 import com.isa.jjdzr.walletcore.service.WalletService;
+import com.isa.jjdzr.walletcore.service.WalletServiceImpl;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
@@ -16,11 +18,11 @@ public class WalletWebConfiguration {
 
     @Bean
     public WalletService walletService() {
-        return new WalletService();
+        return new WalletServiceImpl();
     }
 
     @Bean
     public WalletAssetService walletAssetService() {
-        return new WalletAssetService();
+        return new WalletAssetServiceImpl();
     }
 }
