+++ wallet-web/src/main/java/com/isa/jjdzr/walletweb/controller/WalletController.java
@@ -5,8 +5,6 @@ import com.isa.jjdzr.walletcore.dto.WalletAsset;
 import com.isa.jjdzr.walletcore.service.WalletAssetService;
 import com.isa.jjdzr.walletcore.service.WalletService;
 import com.isa.jjdzr.walletweb.Constants;
-import com.isa.jjdzr.walletweb.dto.User;
-import jakarta.servlet.http.HttpSession;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Controller;
@@ -15,7 +13,6 @@ import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 import java.util.List;
@@ -24,8 +21,8 @@ import java.util.List;
 @RequiredArgsConstructor
 public class WalletController {
 
-    private final WalletService walletService;
-    private final WalletAssetService walletAssetService;
+    private final WalletService walletServiceImpl;
+    private final WalletAssetService walletAssetServiceImpl;
 
     @GetMapping("/create-wallet")
     public String createWallet(Model model) {
@@ -33,22 +30,10 @@ public class WalletController {
         return "create-wallet";
     }
 
-    @PostMapping("/handle-wallet-creation")
-    public String createWallet(@Valid Wallet wallet, BindingResult result, RedirectAttributes redirectAttributes) {
-        if (result.hasErrors()) return "create-wallet";
-        //TODO: add this in thymeleaf (as logged user id)
-        wallet.setUserId(0L);
-        walletService.saveWallet(wallet);
-        String status = Constants.SUCCESS_STATUS;
-        redirectAttributes.addFlashAttribute("status", status);
-        //TODO: adding newly created wallet to session
-        return "redirect:/wallet-view";
-    }
-
     @GetMapping("/wallet-view")
-    public String showWallet(Model model){
-        List<WalletAsset> walletAssets = walletAssetService.findWalletAssetsByWalletId(0L);
-        Wallet wallet = walletService.find(0L);
+    public String showWallet(Model model) {
+        List<WalletAsset> walletAssets = walletAssetServiceImpl.findWalletAssetsByWalletId(0L);
+        Wallet wallet = walletServiceImpl.find(0L);
         model.addAttribute("walletAssets", walletAssets);
         model.addAttribute("wallet", wallet);
         return "wallet-view";
@@ -57,8 +42,20 @@ public class WalletController {
     @GetMapping("/load-wallet/{userId}")
     public String loadWallet(Model model, @PathVariable("userId") Long userId) {
         if (userId == -1L) return "redirect:/login";
-        List<Wallet> walletList = walletService.getUsersWallets(userId);
+        List<Wallet> walletList = walletServiceImpl.getUserWallets(userId);
         model.addAttribute(walletList);
         return "load-wallet";
+    }
+
+    @PostMapping("/handle-wallet-creation")
+    public String createWallet(@Valid Wallet wallet, BindingResult result, RedirectAttributes redirectAttributes) {
+        if (result.hasErrors()) return "create-wallet";
+        //TODO: add this in thymeleaf (as logged user id)
+        wallet.setUserId(0L);
+        walletServiceImpl.saveWallet(wallet);
+        String status = Constants.SUCCESS_STATUS;
+        redirectAttributes.addFlashAttribute("status", status);
+        //TODO: adding newly created wallet to session
+        return "redirect:/wallet-view";
     }
 }
